tester/requirements.txt
requests
pillow
numpy
python-Levenshtein